store intrumented glob in config
