+++ package.json
@@ -12,8 +12,8 @@
         "build": "tsc -p src/lib; tsc -p src/components; tsc -p src/browser; tsc -p src/test; tsc -p src",
         "watch": "tsc -p src/lib -w ",
         "help": "dist/lib/cli.js -h",
-        "test": "dist/lib/cli.js dist/components/*.js -f test/runner.html",
-        "test-dev": "dist/lib/cli.js dist/components/*.js -f test/runner.html -d"
+        "test": "dist/lib/cli.js server-config.json",
+        "test-dev": "dist/lib/cli.js  -f test/runner.html -d"
     },
     "repository": {
         "type": "git",
@@ -29,9 +29,12 @@
     "dependencies": {
         "@hmh/nodejs-base-server": "latest",
         "@types/chai": "latest",
+        "@types/globby": "^8.0.0",
         "@types/mocha": "latest",
+        "D": "^1.0.0",
         "chai": "latest",
         "commander": "latest",
+        "globby": "^8.0.1",
         "istanbul-api": "^2.0.6",
         "istanbul-lib-coverage": "^2.0.1",
         "istanbul-lib-instrument": "^3.0.0",
