+++ src/test/index.spec.ts
@@ -1,22 +1,5 @@
-import { LitElement } from '@polymer/lit-element';
+import pageBannerSpec from './page-banner.spec.js';
 
-const expect: Chai.ExpectStatic = chai.expect;
-
-describe('some example tests', () => {
-    it('should be true', (): void => {
-        expect(true).to.be.true;
-    });
-
-    it('print innerHTML', async (): Promise<void> => {
-        withSnippet('first');
-        const el: LitElement = document.querySelector('page-banner');
-        await el.updateComplete;
-        const message = 'Hello World!';
-        expect(el.shadowRoot).not.to.be.undefined;
-        const h3: HTMLHeadingElement = el.shadowRoot.querySelector('h3');
-
-        expect(h3.innerText).to.equal(message);
-    });
-});
+pageBannerSpec();
 
 mocha.run();
